fix: fixes

+++ api-service/src/model/video.rs
@@ -4,8 +4,8 @@ use ::serde::{Serialize, Deserialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize, sqlx::FromRow)]
 pub struct Video {
-    pub id: i32,
-    pub creator_id: i32,
+    pub id: i64,
+    pub creator_id: i64,
     pub name: String,
     pub preview_url: String,
     pub video_url: String,
